backend/requirements.txt
Django>=5.2
psycopg2-binary>=2.9
djangorestframework>=3.15
django-cors-headers>=4.3
python-dotenv>=1.0
gunicorn>=21.2.0
djangorestframework-simplejwt>=5.3.1